+++ src/headers.ts
@@ -147,6 +147,7 @@ export type EightyNumberTuple = [
   ...TenNumberTuple,
   ...TenNumberTuple
 ];
+export type WarningFlagsTuple = [[number, number], [number, number]];
 
 /*************************************************
  *
@@ -598,10 +599,25 @@ export const inchi_Input = NAPIStructType({
   num_stereo0D: refNAPI.types.short,
 });
 export interface INCHIInput {
+  /**
+   * Array of Inchi Atoms
+   */
   atom: INCHIAtom[];
-  stereo0D: INCHIStereo0D[];
+  /**
+   * Array of num_stereo0D 0D stereo elements or NULL
+   */
+  stereo0D: INCHIStereo0D[] | null;
+  /**
+   * InChI options: space-delimited; each is preceded by '/' or '-' depending on OS and compiler
+   */
   szOptions: string;
+  /**
+   * number of atoms in the structure < MAX_ATOMS
+   */
   numAtoms: number;
+  /**
+   * number of 0D stereo elements
+   */
   numStereo0D: number;
 }
 
@@ -650,22 +666,58 @@ export const inchi_Input_PolymerUnit = NAPIStructType({
   xbr1: NAPIArrayType(refNAPI.types.double, 4),
   xbr2: NAPIArrayType(refNAPI.types.double, 4),
   smt: NAPIArrayType(refNAPI.types.char, 80),
-  alist: refNAPI.refType(refNAPI.types.int),
-  blist: refNAPI.refType(refNAPI.types.int),
+  alist: NAPIArrayType(refNAPI.types.int),
+  blist: NAPIArrayType(refNAPI.types.int),
 });
 export interface INCHIInputPolymerUnit {
+  /**
+   * Unit id , it is what is called 'Sgroup number'  in CTFile (not used, kept for compatibility)
+   */
   id: number;
+  /**
+   * Unit type as per CTFile format (STY)
+   */
   type: number;
+  /**
+   * Unit subtype as per CTFile format (SST)
+   */
   subType: number;
+  /**
+   * Unit connection scheme  as per CTFile format (SCN)
+   */
   conn: number;
+  /**
+   * One more unit id; what is called 'unique Sgroup identifier' in CTFile (not used, for compatibility)
+   */
   label: number;
+  /**
+   * Number of atoms in the unit
+   */
   nA: number;
+  /**
+   * Number of bonds in the unit
+   */
   nB: number;
+  /**
+   *  Bracket ends coordinates (SDI)
+   */
   xBr1: FourNumberTuple;
+  /**
+   *  Bracket ends coordinates (SDI)
+   */
   xBr2: FourNumberTuple;
+  /**
+   * Sgroup Subscript (SMT) ('n' or so )
+   */
   smt: EightyNumberTuple;
-  aList: number;
-  bList: number;
+  /**
+   * List of atoms in the unit (SAL), atomic numbers
+   */
+  aList: number[];
+  /**
+   * List of crossing bonds of unit: [bond1end1, bond1end2, bond2end1, bond2end2]
+   */
+  bList: number[];
 }
 
 // typedef struct inchi_Input_Polymer
@@ -676,11 +728,17 @@ export interface INCHIInputPolymerUnit {
 // } inchi_Input_Polymer;
 
 export const inchi_Input_Polymer = NAPIStructType({
-  units: refNAPI.refType(refNAPI.refType(inchi_Input_PolymerUnit)),
+  units: NAPIArrayType(refNAPI.refType(inchi_Input_PolymerUnit), 999),
   n: refNAPI.types.int,
 });
 export interface INCHIInputPolymer {
+  /**
+   * List of pointers to polymer units
+   */
   units: INCHIInputPolymerUnit[];
+  /**
+   * Number of polymer units
+   */
   n: number;
 }
 
@@ -730,30 +788,61 @@ export const inchi_Input_V3000 = NAPIStructType({
   n_collections: refNAPI.types.int,
   n_non_haptic_bonds: refNAPI.types.int,
   n_haptic_bonds: refNAPI.types.int,
-  lists_haptic_bonds: refNAPI.refType(refNAPI.refType(refNAPI.types.int)),
+  lists_haptic_bonds: NAPIArrayType(refNAPI.types.int, 999),
   n_steabs: refNAPI.types.int,
-  lists_steabs: refNAPI.refType(refNAPI.refType(refNAPI.types.int)),
+  lists_steabs: NAPIArrayType(NAPIArrayType(refNAPI.types.int, 999), 999),
   n_sterel: refNAPI.types.int,
-  lists_sterel: refNAPI.refType(refNAPI.refType(refNAPI.types.int)),
+  lists_sterel: NAPIArrayType(NAPIArrayType(refNAPI.types.int, 999), 999),
   n_sterac: refNAPI.types.int,
-  lists_sterac: refNAPI.refType(refNAPI.refType(refNAPI.types.int)),
+  lists_sterac: NAPIArrayType(NAPIArrayType(refNAPI.types.int, 999), 999),
 });
 export interface INCHIInputV3000 {
   nNonStartAtoms: number;
   nStarAtoms: number;
+  /**
+   * Array of Index of atoms
+   */
   atomIndexOrig: number;
+  /**
+   *  Array of index ,=index or -1 for star atom
+   */
   atomIndexFin: number;
+  /**
+   * Not used yet.
+   */
   nSGroups: number;
+  /**
+   * Not used yet.
+   */
   n3DContraints: number;
   nCollections: number;
   nNonHapticBonds: number;
   nHapticBonds: number;
-  listsHapticBonds: number[][];
+  /**
+   * Haptic_bonds[i] is pointer to int array which contains:
+   * bond_type, non-star atom number, nendpts, then endpts themselves
+   */
+  listsHapticBonds: number[];
   nSteabs: number;
+  /**
+   * steabs[k][0] - not used
+   * steabs[k][1] -  number of members in collection
+   * steabs[k][2..] - member atom numbers
+   */
   listsSteabs: number[][];
   nSterel: number;
+  /**
+   * sterel[k][0] - n from "STERELn" tag
+   * sterel[k][1] -  number of members in collection
+   * sterel[k][2..] - member atom numbers
+   */
   listsSterel: number[][];
   nSterac: number;
+  /**
+   * sterac[k][0] - n from "STERACn" tag
+   * sterac[k][1] -  number of members in collection
+   * sterac[k][0] - number from "STERACn" tag
+   */
   listsSterac: number[][];
 }
 
@@ -781,8 +870,8 @@ export interface INCHIInputV3000 {
 // } inchi_InputEx;
 
 export const inchi_InputEx = NAPIStructType({
-  atom: NAPIArrayType(inchi_Atom),
-  Stereo0D: NAPIArrayType(inchi_Stereo0D),
+  atom: NAPIArrayType(inchi_Atom, 999),
+  Stereo0D: NAPIArrayType(inchi_Stereo0D, 999),
   szOptions: refNAPI.types.CString,
   num_atoms: refNAPI.types.short,
   num_stereo0D: refNAPI.types.short,
@@ -790,13 +879,34 @@ export const inchi_InputEx = NAPIStructType({
   v3000: NAPIArrayType(inchi_Input_V3000),
 });
 export interface INCHIInputEx {
+  /**
+   * Array of INCHIAtom elements
+   */
   atom: INCHIAtom[];
-  stereo0D: INCHIStereo0D[];
+  /**
+   * Array of INCHIstereo0D 0D stereo elements
+   */
+  stereo0D: INCHIStereo0D[] | null;
+  /**
+   * InChI options: space-delimited; each is preceded by '/' or '-' depending on OS and compiler
+   */
   szOptions: string | null;
+  /**
+   * number of atoms in the structure
+   */
   numAtoms: number;
+  /**
+   * number of 0D stereo element, NULL if not a polymer
+   */
   numStereo0D: number;
-  polymer: INCHIInputPolymer[];
-  v3000: INCHIInputV3000[];
+  /**
+   * v. 1.05+ extended data, polymers
+   */
+  polymer: INCHIInputPolymer | null;
+  /**
+   * v. 1.05+ extended data, V3000 Molfile features, NULL if no V3000 extensions present
+   */
+  v3000: INCHIInputV3000 | null;
 }
 
 // /*
@@ -819,7 +929,13 @@ export const inchi_InputINCHI = NAPIStructType({
   szOptions: refNAPI.types.CString,
 });
 export interface INCHIInputINCHI {
+  /**
+   * InChI ASCIIZ string to be converted to a strucure
+   */
   szINCHI: string | null;
+  /**
+   * InChI options: space-delimited; each is preceded by '/' or '-' depending on OS and compiler
+   */
   szOptions: string | null;
 }
 
@@ -857,9 +973,21 @@ export const inchi_Output = NAPIStructType({
   szLog: refNAPI.types.CString,
 });
 export interface INCHIOutput {
+  /**
+   * InChI ASCIIZ string
+   */
   szINCHI: string | null;
+  /**
+   * Aux info ASCIIZ string
+   */
   szAuxInfo: string | null;
+  /**
+   * Error/warning ASCIIZ message
+   */
   szMessage: string | null;
+  /**
+   * log-file ASCIIZ string, contains a human-readable list of recognized options and possibly an Error/warning message
+   */
   szLog: string | null;
 }
 
@@ -886,8 +1014,8 @@ export interface INCHIOutput {
 // }inchi_OutputStruct;
 
 export const inchi_OutputStruct = NAPIStructType({
-  atom: refNAPI.refType(inchi_Atom),
-  stereo0D: refNAPI.refType(inchi_Stereo0D),
+  atom: NAPIArrayType(inchi_Atom),
+  stereo0D: NAPIArrayType(inchi_Stereo0D),
   num_atoms: refNAPI.types.short,
   num_stereo0D: refNAPI.types.short,
   szMessage: refNAPI.types.CString,
@@ -895,13 +1023,38 @@ export const inchi_OutputStruct = NAPIStructType({
   WarningFlags: NAPIArrayType(NAPIArrayType(refNAPI.types.ulong, 2), 2),
 });
 export interface INCHIOutputStruct {
-  atom: INCHIAtom;
-  stereo0D: INCHIStereo0D;
+  /**
+   * Array of INCHIAtom elements
+   */
+  atom: INCHIAtom[];
+  /**
+   * Array of INCHIStereo0D 0D stereo elements or NULL
+   */
+  stereo0D: INCHIStereo0D[] | null;
+  /**
+   * number of atoms in the structure
+   */
   numAtoms: number;
+  /**
+   * number of 0D stereo elements
+   */
   numStereo0D: number;
+  /**
+   * Error/warning ASCIIZ message
+   */
   szMessage: string | null;
+  /**
+   * log-file ASCIIZ string, contains a human-readable list of recognized options and possibly an Error/warn message
+   */
   szLog: string | null;
-  warningFlags: [[number, number], [number, number]];
+  /**
+   * warnings [x][y]:
+   * - `x=0` => Reconnected if present in InChI otherwise Disconnected/Normal
+   * - `x=1` => Disconnected layer if Reconnected layer is present
+   * - `y=0` => Fixed-H layer
+   * - `y=1` => Main layer or Mobile-H
+   */
+  warningFlags: WarningFlagsTuple;
 }
 
 // typedef struct tagINCHI_OutputStructEx
@@ -926,8 +1079,8 @@ export interface INCHIOutputStruct {
 // } inchi_OutputStructEx;
 
 export const inchi_OutputStructEx = NAPIStructType({
-  atom: refNAPI.refType(inchi_Atom),
-  stereo0D: refNAPI.refType(inchi_Stereo0D),
+  atom: NAPIArrayType(inchi_Atom),
+  stereo0D: NAPIArrayType(inchi_Stereo0D),
   num_atoms: refNAPI.types.short,
   num_stereo0D: refNAPI.types.short,
   szMessage: refNAPI.types.CString,
@@ -937,15 +1090,15 @@ export const inchi_OutputStructEx = NAPIStructType({
   v3000: refNAPI.refType(inchi_Input_V3000),
 });
 export interface INCHIOutputStructEx {
-  atom: INCHIAtom;
-  stereo0D: INCHIStereo0D;
+  atom: INCHIAtom[];
+  stereo0D: INCHIStereo0D[] | null;
   numAtoms: number;
   numStereo0D: number;
   szMessage: string | null;
   szLog: string | null;
-  warningFlags: [[number, number], [number, number]];
-  polymer: INCHIInputPolymer;
-  v3000: INCHIInputV3000;
+  warningFlags: WarningFlagsTuple;
+  polymer: INCHIInputPolymer | null;
+  v3000: INCHIInputV3000 | null;
 }
 
 // void FreeInChIExtInput( inchi_Input_Polymer    *polymer, inchi_Input_V3000 *v3000 );